photo_selector/requirements.txt
# Photo Selection Pipeline Requirements

# Core ML
torch>=2.0.0
torchvision>=0.15.0

# CLIP for embeddings
# Install with: pip install git+https://github.com/openai/CLIP.git
# Or use: pip install openai-clip

# Image processing
Pillow>=10.0.0
pillow-heif>=0.13.0  # HEIC support
opencv-python>=4.8.0

# Machine learning
numpy>=1.24.0
scikit-learn>=1.3.0

# Optional but recommended
tqdm>=4.65.0  # Progress bars
